requirements.txt
aerich==0.7.2
fastapi==0.115.0
uvicorn[standard]==0.32.0
tortoise-orm==0.20.0
aerich==0.7.2
asyncpg==0.29.0
dramatiq==1.16.0
redis==5.0.1
pydantic==2.8.2
pydantic-settings==2.5.2
aiogram==3.4.1
python-dotenv==1.0.1
openai==1.51.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
